Update vmware-comp-guide.md
Updated Line 47 to correct what TPM stands for just to see if I can make changes and if they will be accepted. I am a technical writer for IBM Cloud and there are a few issues with this document I would like to update. Feel free to contact me over slack.